Refactor Document Generator: Remove technical implementation and visual design documents, streamline UI components, and enhance testing coverage
- Deleted technical implementation and visual design mockup documents to simplify project structure.
- Refactored UI components to create resource and section editors with improved structure and clarity.
- Enhanced test coverage for outline models, UI components, and validation logic.
- Updated documentation server to support both file paths and URLs, improving flexibility in handling documentation sources.
- Adjusted example configuration to reflect new documentation source.
