--- apps/document-generator/document_generator_app/cli/__main__.py
@@ -1,8 +1,0 @@
-"""
-CLI entrypoint for Document Generator.
-"""
-
-from .main import app
-
-if __name__ == "__main__":
-    app()
