--- apps/document-generator/document_generator_app/config/__init__.py
@@ -1,6 +1,0 @@
-"""
-Config package for Document Generator.
-"""
-
-__all__ = ["Settings"]
-from .settings import Settings
