--- apps/document-generator/document_generator_app/schema/__init__.py
@@ -1,6 +1,0 @@
-"""
-Schema package for Document Generator.
-"""
-
-__all__ = ["validate_outline"]
-from .validator import validate_outline
