--- apps/document-generator/document_generator_app/schema/outline_schema.json
@@ -1,47 +1,0 @@
-{
-  "$schema": "http://json-schema.org/draft-07/schema#",
-  "title": "Outline",
-  "type": "object",
-  "properties": {
-    "title": {"type": "string"},
-    "general_instruction": {"type": "string"},
-    "resources": {
-      "type": "array",
-      "items": {
-        "type": "object",
-        "properties": {
-          "key": {"type": "string"},
-          "path": {"type": "string"},
-          "description": {"type": "string"},
-          "merge_mode": {"type": "string", "enum": ["concat", "dict"]}
-        },
-        "required": ["key", "path", "description"],
-        "additionalProperties": false
-      }
-    },
-    "sections": {
-      "type": "array",
-      "items": {"$ref": "#/definitions/section"}
-    }
-  },
-  "definitions": {
-    "section": {
-      "type": "object",
-      "properties": {
-        "title": {"type": "string"},
-        "prompt": {"type": "string"},
-        "refs": {"type": "array", "items": {"type": "string"}},
-        "resource_key": {"type": "string"},
-        "sections": {"type": "array", "items": {"$ref": "#/definitions/section"}}
-      },
-      "required": ["title"],
-      "oneOf": [
-        {"required": ["prompt"]},
-        {"required": ["resource_key"]}
-      ],
-      "additionalProperties": false
-    }
-  },
-  "required": ["title", "general_instruction", "resources", "sections"],
-  "additionalProperties": false
-}
