--- apps/document-generator/document_generator_app/ui/__init__.py
@@ -1,6 +1,0 @@
-"""
-Simplified Document Generator UI package.
-"""
-from document_generator_app.ui.layout import build_editor
-
-__all__ = ["build_editor"]
