--- apps/document-generator/plans/IMPLEMENTATION_PROGRESS.md
@@ -1,114 +1,0 @@
-# Document Generator UI Simplification - Implementation Progress
-
-## Overview
-Tracking progress on simplifying the Document Generator UI from complex dropdowns/callbacks to clean two-column direct-selection interface.
-
-## Key Design Decisions (Locked In)
-- **Layout**: Two columns - list management left, editor right
-- **Selection**: Direct click to select (no dropdowns)
-- **Subsections**: "+Sub" button, max 4 levels, visual indentation
-- **State**: Minimal - just track selected item
-- **Callbacks**: Only 5 total (select, add_resource, add_section, save_item, remove_selected)
-- **Migration**: mv ui/ ui_old/, build new in ui/, no backwards compat
-
-## Implementation Checklist
-
-### Phase 1: Setup ✅
-- [x] Move old UI: `mv ui/ ui_old/`
-- [x] Create new UI structure
-- [x] Copy reusable utilities from ui_old/utils.py
-
-### Phase 2: Core Components ✅
-- [x] ui/components.py
-  - [x] create_resource_list()
-  - [x] create_section_list() 
-  - [x] create_resource_editor()
-  - [x] create_section_editor()
-  - [x] render_resource_items()
-  - [x] render_section_items() with nesting
-- [x] ui/callbacks.py
-  - [x] select_item()
-  - [x] add_resource()
-  - [x] add_section(as_subsection=False)
-  - [x] save_item()
-  - [x] remove_selected()
-- [x] ui/layout.py
-  - [x] Two-column layout
-  - [x] Wire up all callbacks
-  - [x] Empty state handling
-
-### Phase 3: Integration ⬜
-- [ ] Connect to existing executor
-- [ ] File upload/download
-- [ ] Generate document functionality
-- [ ] Test all features work
-
-### Phase 4: Cleanup ⬜
-- [ ] Remove ui_old/
-- [ ] Update any imports
-- [ ] Update main.py if needed
-
-## Current Status
-**Stage**: Radio Button UI Successfully Implemented!
-**Next Step**: Test all features work end-to-end with the new UI
-
-## Issues Fixed
-- Removed emoji characters that were causing encoding issues
-- Fixed nested f-string syntax error in components.py
-- Verified all imports work correctly
-- Fixed file upload: Gradio returns file path, not file object
-- Fixed dataclass vs dictionary mismatch: Updated all UI code to work with Section/Resource dataclasses
-- **MAJOR FIX**: Replaced hacky HTML/JavaScript selection with native Gradio Radio components
-  - No more JavaScript event handling
-  - No more HTML generation
-  - Clean, simple radio button selection
-  - Subsections shown with text indentation (└─ prefix)
-  - Following "ruthless simplicity" philosophy
-- **AUTO-SAVE**: Implemented automatic saving without Save buttons
-  - Changes save immediately as you type
-  - Radio labels update in real-time (e.g., resource key changes)
-  - No UI flashing or feedback loops
-  - Much better user experience
-- **RESOURCE KEY CHANGES**: Simplified approach following "ruthless simplicity"
-  - When resource keys change, invalid refs are filtered at display time
-  - No complex cascading updates or state synchronization
-  - Clear UI note about needing to re-select sections after key changes
-  - Clears section selection when resource key changes to avoid conflicts
-  - Avoids hacky workarounds and complex error handling
-
-## How to Run
-```bash
-# From the root directory with venv activated:
-document-generator-app
-```
-
-## Key Code Patterns to Remember
-
-### State Structure
-```python
-state = {
-    "outline": Outline(),
-    "selected_type": None,  # "resource" or "section"
-    "selected_id": None,    # e.g., "resource_0" or "section_1_2"
-}
-```
-
-### Section Path Handling
-```python
-# section_1_2_0 means: sections[1].sections[2].sections[0]
-path = selected_id.split("_")[1:]  # ["1", "2", "0"]
-```
-
-### Subsection Addition
-```python
-if as_subsection and state["selected_type"] == "section":
-    # Add under selected section (if < 4 levels deep)
-else:
-    # Add as top-level section
-```
-
-## Notes for Future Sessions
-- All plans in /apps/document-generator/plans/
-- Focus on simplicity - no fancy features
-- Test manually first, automate later
-- Reuse existing models/executor/validator
