--- apps/document-generator/tests/config/test_settings.py
@@ -1,8 +1,0 @@
-from document_generator_app.config.settings import Settings
-
-
-def test_settings_defaults(monkeypatch):
-    # Should pick up environment override
-    monkeypatch.setenv("MODEL_NAME", "test-model")
-    s = Settings()
-    assert s.model_name == "test-model"
